Add basic integration test; Move core function Added a basic integration test to make sure everything runs as expected.
Moved the main extract_metadata() function to __init__
so it can be imported.
